requirements.txt
flask
openai
pandas
langchain
langchain-community
langchain-experimental
bcrypt
python-dotenv
